# swagger.yaml
openapi: 3.0.3
info:
  title: API de Monitoreo de Calidad del Aire
  description: API REST para el sistema de monitoreo de calidad del aire con arquitectura hexagonal
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: soporte@monitoreo-aire.com
servers:
  - url: http://localhost:5000/api
    description: Servidor de desarrollo
security:
  - bearerAuth: []
  - deviceToken: []
paths:
...
  /users:
    get:
      summary: Obtener todos los usuarios
      description: Lista todos los usuarios del sistema (solo administradores)
      tags:
        - Usuarios
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
        401:
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Permisos insuficientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /devices:
    post:
      summary: Crear dispositivo
      description: Registra un nuevo dispositivo de monitoreo (solo administradores)
      tags:
        - Dispositivos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
                - ubicacion
                - id_edificio
              properties:
                nombre:
                  type: string
                  example: "Sensor Piso 1 - Sala A"
                ubicacion:
                  type: string
                  example: "Edificio Principal, Piso 1, Sala de Reuniones A"
                id_edificio:
                  type: integer
                  example: 1
      responses:
        201:
          description: Dispositivo creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispositivo'
        403:
          description: Permisos insuficientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Obtener dispositivos
      description: Lista todos los dispositivos del sistema
      tags:
        - Dispositivos
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de dispositivos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dispositivo'
...
components:
  schemas:
    Usuario:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Juan Pérez"
        correo:
          type: string
          format: email
          example: "juan.perez@email.com"
        rol:
          type: string
          enum: [administrador, conserje]
          example: "conserje"
        numero_contacto:
          type: string
          example: "+56912345678"
        url_imagen_perfil:
          type: string
          format: uri
          example: "https://example.com/avatar.jpg"
        id_edificio_asignado:
          type: integer
          example: 1
        fecha_creacion:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
    Dispositivo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: "Sensor Piso 1 - Sala A"
        ubicacion:
          type: string
          example: "Edificio Principal, Piso 1, Sala de Reuniones A"
        token_acceso:
          type: string
          example: "abc123def456ghi789"
        id_edificio:
          type: integer
          example: 1
        estado:
          type: string
          enum: [activo, inactivo, mantenimiento]
          example: "activo"
        fecha_instalacion:
          type: string
          format: date
          example: "2025-01-15"
        ultima_lectura:
          type: string
          format: date-time
          example: "2025-01-15T14:30:00Z"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Mensaje de error descriptivo"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /login

    deviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: Token único del dispositivo para autenticación de IoT